Update upgrate path for new core tools

+++ admin/lti/database.php
@@ -839,9 +839,12 @@ $DATABASE_UPGRADE = function($oldversion) {
     $indexes = $PDOX->indexes($CFG->dbprefix."lti_key");
 
     // DROP INDEX index_name ON tbl_name
+    $keep_issuer_index = $PDOX->columnExists('issuer_id', $CFG->dbprefix."lti_key");
     foreach($indexes as $index) {
         if ( strcasecmp($index, "PRIMARY") == 0 ) continue;
         if ( strcasecmp($index, "ibfk") == 0 ) continue;
+        // lti_key_const_2 backs issuer_id until lti_issuer phase 3 (2026-10-01 UTC)
+        if ( $keep_issuer_index && $index === $CFG->dbprefix."lti_key_const_2" ) continue;
         $command = isset($needed_indexes[$index]) ? $needed_indexes[$index] : null;
         if ( is_string($command) ) continue;
         $sql = "DROP INDEX ".$index." ON ".$CFG->dbprefix."lti_key";
